PatronTransaction: Allow linking to a Loan
At this time, all PatronTransaction are either linked to a Notification,
either linked to a Patron (for subscription). But the future incremental
fees will not be linked to a Notification but need to be linked to a
Loan. Updates the JSON schema to allow a PatronTransaction to be
directly linked to a Loan resource.

Co-Authored-by: Renaud Michotte <renaud.michotte@gmail.com>